+++ backend/src/main/java/com/pobluesky/backend/domain/user/service/ManagerService.java
@@ -7,15 +7,10 @@ import com.pobluesky.backend.domain.user.entity.Manager;
 import com.pobluesky.backend.domain.user.repository.ManagerRepository;
 import com.pobluesky.backend.global.error.CommonException;
 import com.pobluesky.backend.global.error.ErrorCode;
-import com.pobluesky.backend.global.security.JwtToken;
-import com.pobluesky.backend.global.security.JwtTokenProvider;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -31,36 +26,18 @@ public class ManagerService {
 
     private final ManagerRepository managerRepository;
 
-    private final AuthenticationManagerBuilder authenticationManagerBuilder;
-
-    private final JwtTokenProvider jwtTokenProvider;
-
     private final PasswordEncoder passwordEncoder;
-
-    @Transactional
-    public JwtToken signIn(String email, String password) {
-        UsernamePasswordAuthenticationToken authenticationToken =
-            new UsernamePasswordAuthenticationToken(email, password);
-
-        Authentication authentication = authenticationManagerBuilder
-            .getObject()
-            .authenticate(authenticationToken);
-
-        JwtToken jwtToken = jwtTokenProvider.generateToken(authentication);
-
-        return jwtToken;
-    }
 
     @Transactional
     public ManagerResponseDTO signUp(ManagerCreateRequestDTO signUpDto) {
         if (managerRepository.existsByEmail(signUpDto.email())) {
-            throw new IllegalArgumentException("이미 사용 중인 이메일 입니다.");
+            throw new CommonException(ErrorCode.ALREADY_EXISTS_EMAIL);
         }
 
-        // Password 암호화
         String encodedPassword = passwordEncoder.encode(signUpDto.password());
+
         List<String> roles = new ArrayList<>();
-        roles.add("USER");  // USER 권한 부여
+        roles.add("USER");
 
         Manager manager = signUpDto.toManagerEntity(
             encodedPassword,
@@ -78,14 +55,6 @@ public class ManagerService {
             .map(ManagerResponseDTO::from)
             .collect(Collectors.toList());
     }
-
-//    @Transactional
-//    public ManagerResponseDTO createManager(ManagerCreateRequestDTO dto) {
-//        Manager manager = dto.toManagerEntity();
-//        Manager savedManager = managerRepository.save(manager);
-//
-//        return ManagerResponseDTO.from(savedManager);
-//    }
 
     @Transactional
     public ManagerResponseDTO updateManagerById(
@@ -109,5 +78,4 @@ public class ManagerService {
     public void deleteManagerById(Long userId) {
         managerRepository.deleteById(userId);
     }
-
 }
